+++ src/lib.rs
@@ -21,6 +21,8 @@
 //! The client-driver is distributed under the liberal [Apache-2.0 License](https://www.apache.org/licenses/LICENSE-2.0) and hence
 //! you can use it in your applications without any licensing issues.
 //!
+//! **📁 You can [find some usage examples in this folder here](https://github.com/skytable/client-rust/tree/v0.8.11/examples)**.
+//!
 //! ## Definitive example
 //!
 //! ```no_run
@@ -35,7 +37,7 @@
 //!     userid: String,
 //!     pass: String,
 //!     followers: u64,
-//!     email: Option<String>
+//!     email: Option<String>,
 //! }
 //!
 //! let our_user = User { userid: "user".into(), pass: "pass".into(), followers: 120, email: None };